删除分类Bug修复，现在不能删除已经关联了菜品或者套餐的分类了

+++ sky-server/src/main/java/com/sky/controller/admin/CategoryController.java
@@ -63,7 +63,7 @@ public class CategoryController {
 
     @DeleteMapping
     @ApiOperation("跟据ID删除分类")
-    public Result deletById(Integer id) {
+    public Result deletById(Long id) {
         log.info("跟据ID删除分类：{}", id);
         categoryService.deleteById(id);
         return Result.success();
